+++ common/kernels/imageBlur_kernel_stage2.cu
@@ -36,7 +36,7 @@ __global__ void blurKernel_st2(float* outputPixels,float* intermediate, int* wei
         j *= yScale;
     }
     float kernelSum = 16.0;
-    extern __shared__ float ghb[]; // for per-thread local history
+    extern __shared__ int ghb[]; // for per-thread local history
     int my_ghb_index = ((threadIdx.y * blockDim.x) + threadIdx.x) * 3;
 
     // still check for this in case of small img, not all threads need execute
@@ -65,9 +65,9 @@ __global__ void blurKernel_st2(float* outputPixels,float* intermediate, int* wei
                   // finish up last few values with NUM_TEX reads
                   for(int ii = (RADIUS-NUM_TEX+1); ii <= RADIUS;ii++) {
                       int filterWeightLoc = RADIUS + ii;
-                      float curValueHash = hashGHB(&ghb[my_ghb_index]);
-                      float texVal = tex1D<float>(tref,curValueHash);
-                      tmp += (ghb[my_ghb_index+2] + texVal) * weightedKernel[filterWeightLoc];
+                      int curValueHash = hashGHB(&ghb[my_ghb_index]);
+                      int texVal = tex1D<int>(tref,curValueHash);
+                      tmp += (float)(ghb[my_ghb_index+2] + texVal) * weightedKernel[filterWeightLoc];
                   }
                 float avg = tmp / kernelSum;
                 outputPixels[curElement] = avg;
