+++ common/kernels/imageResize_kernel_st2.cu
@@ -39,7 +39,7 @@ __global__ void resizeKernel_st2(float* outputPixels,float* intermediate, int* w
     }
 
     float kernelSum = 16.0;
-    extern __shared__ float ghb[]; // for per-thread local history
+    extern __shared__ int ghb[]; // for per-thread local history
     int my_ghb_index = ((threadIdx.y * blockDim.x) + threadIdx.x) * 3;
     
     for (int idx = i; idx < (i + xScale); idx++) {
@@ -67,9 +67,9 @@ __global__ void resizeKernel_st2(float* outputPixels,float* intermediate, int* w
 
                 // finish up last few values with NUM_TEX reads
                 for(int ii = (DIAMETER-NUM_TEX); ii < DIAMETER;ii++) {
-                    float curValueHash = hashGHB(&ghb[my_ghb_index]);
-                    float texVal = tex1D<float>(tref,curValueHash);
-                    tmp += (ghb[my_ghb_index+2] + texVal) * weightedKernel[ii];
+                    int curValueHash = hashGHB(&ghb[my_ghb_index]);
+                    int texVal = tex1D<int>(tref,curValueHash);
+                    tmp += (float)(ghb[my_ghb_index+2] + texVal) * weightedKernel[ii];
                 }
                 float avg = tmp / kernelSum;
                 outputPixels[elemToWrite] = avg;
